Move WebauthnListener into the Gem::GemcutterUtilities namespace
